utils (hyperscript, pageinfo) moved to a package
