+++ customblocks/utils/__init__.py
@@ -1,0 +1,2 @@
+from .hyperscript import E, Markdown
+from .pageinfo import PageInfo
